diffraction: extract estimate_peak_windows
Appended to peak_detection.py, which now covers the detection path end to end:
model input prep, blob detection and refinement, output post-processing, and the
radial families the peaks cluster into. Two attribute reads threaded in as leading
parameters, both documented in the existing numpydoc block; the class keeps a
forwarding method since 51 notebooks call it.

Verified per protocol: baseline tests first (90cf2cb) -- it had none, despite
deciding the q-windows every count map, radial profile and flowline family
downstream is computed over. Then AST scan for injected-parameter shadowing
(none), then refactor_diff_check.py: 33/33 bit-identical, with seven new checks
covering the returned centres, windows, and five peak_info arrays separately.

BraggPeaksPolymer: 3193 -> 3050 lines. bragg_peaks.py 3663 -> 3521.

Tests: 511 passed, 10 skipped.
